Добавлены дробные числа

+++ Exemple004/Program.cs
@@ -1,3 +1,3 @@
-﻿int namber_a = 20;
-int namber_b = 15;
+﻿double namber_a = 20;
+double namber_b = 15;
 Console.WriteLine(namber_a / namber_b);
